requirements.txt
streamlit
python-pptx
requests
torch
diffusers
transformers
accelerate
